@@ -15,15 +15,15 @@
 		}
 	],
 	"require": {
-		"php": ">= 7.1.19",
+		"php": ">= 7.2.0",
 		"nette/nette": "~2.4.0",
 		"dg/adminer-custom": "~1.5",
 		"kdyby/forms-replicator": "@dev",
-		"o5/grido": "dev-master",
+		"o5/grido": "~3.0.0",
 		"mpdf/mpdf": ">= 5.6.5",
-		"jkuchar/pdfresponse": "dev-master",
+		"jkuchar/pdfresponse": "~0.4.0-beta2",
 	  	"pear2/net_transmitter": ">=1.0.0a4",
-		"pear2/net_routeros": "dev-master",
+		"pear2/net_routeros": "~1.0.0b6",
 		"nextras/migrations": "^3.0",
 		"guzzlehttp/guzzle": "~6.0",
 		"pauluswebster/shutterstock-rickshaw": "dev-master",
